pyproject.toml
[project]
name = "quero-cafe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-fasthtml>=0.12.31",
]
